2.0.1

+++ src/product.cpp
@@ -1,4 +1,5 @@
 #include "product.h"
+#include <math.h>
 #include <QString>
 
 Product::Product()
@@ -122,7 +123,7 @@ int Product::getPakNumber () const
 {
     if (m_pakUnit == 0)
         return 0;
-    return (int)ceilf((float)m_number / (float)m_pakUnit);
+    return (int)ceil((float)m_number / (float)m_pakUnit);
 }
 QString Product::id() const
 {
